upadate

+++ src/interfaces/Asset.java
@@ -15,7 +15,7 @@ public interface Asset extends Serializable {
     //return an array to store the information of the asset,arr[0] is the rent price, arr[1] is the sell price,
     // arr[2] is the cost price,
     double[] getPrice();
-    void setPrice(int rentPrice,int sellPrice,int costPrice);
+    void setPrice(double rentPrice,double sellPrice,double costPrice);
     Person getOwner();
     void setOwner(Person owner);
     Person getUser();
